Update timeseries forecast via websocket
