+++ lib/phoenix_app/timeseries.ex
@@ -1,69 +1,30 @@
-defmodule PhoenixApp.Timeseries do
+defmodule PhoenixApp.TimeSeries do
   use GenServer
-  use Timex
 
-  @twitter_format "%a %b %d %H:%M:%S %z %Y"
+  require Logger
 
   def start_link do
     GenServer.start_link(__MODULE__, %{}, name: __MODULE__)
   end
 
-  def init(state) do
-    schedule_work()
-    { :ok, state }
+  def init(%{}) do
+    send self(), :timeseries
+    {:ok, %{}}
   end
 
-  def handle_info(:work, state) do
-    listen()
-    { :noreply, state }
+  def handle_info(:timeseries, state) do
+    timeseries()
+    {:noreply, state}
   end
 
-  def schedule_work do
-    delay = 1 * 1000
-    Process.send_after(self(), :work, delay)
-  end
-
-  def listen do
-    rabbitmq = System.get_env("RABBITMQ_HOST")
-    { :ok, connection } = AMQP.Connection.open "amqp://guest:guest@#{rabbitmq}"
-    { :ok, channel } = AMQP.Channel.open connection
-    AMQP.Queue.declare channel, "tweets"
-    AMQP.Basic.consume channel, "tweets", nil, no_ack: true
-
-    wait_for_messages()
-  end
-
-  def wait_for_messages do
-    # NOTE: This receive block should be refactored into the handle_info method.
-    # In the GenServer docs, it explains that the point of GenServer is to
-    # abstract away the need for explicit receive statements.
-    receive do
-      { :basic_deliver, payload, _meta } ->
-        { :ok, tweet } = Poison.decode payload
-
-        count_tweet tweet
-
-        PhoenixAppWeb.Endpoint.broadcast! "twitter:stream", "tweet",
-          %{ body: tweet["text"] }
-
-        wait_for_messages()
+  def timeseries() do
+    case HTTPoison.get("http://timeseries.datapun.net:8003") do
+      {:ok, %HTTPoison.Response{ body: body }} ->
+        response = Poison.decode!(body)
+        PhoenixAppWeb.Endpoint.broadcast!("twitter:stream", "timeseries", response)
+      {:error, %HTTPoison.Error{reason: reason}} -> Logger.warn reason
     end
-  end
 
-  def count_tweet tweet do
-    { :ok, time } = Timex.parse tweet["created_at"], @twitter_format, :strftime
-
-    day = %{ time | hour: 0, minute: 0, second: 0 }
-    minute = %{ time | second: 0 }
-
-    hash = "#{tweet["lang"]}:#{Timex.to_unix day}"
-    key = Timex.to_unix minute
-
-    { :ok, reply } = Redix.command :redix, ~w(hincrby #{hash} #{key} 1)
-
-    if reply == 1 do
-      expiry = Timex.to_unix(day) + (30 * 24 * 60 * 60)
-      { :ok, _ } = Redix.command :redix, ~w(expireat #{hash} #{expiry})
-    end
+    Process.send_after self(), :timeseries, 500
   end
 end
